Separar las clases en ficheros
